Package refactor | Sending product id as parameter
