+++ src/App.jsx
@@ -2,8 +2,9 @@ import logo from './logo.svg';
 import './App.css';
 import { BrowserRouter as Router, Route, Routes, Navigate } from 'react-router-dom';
 import LoginReg from './components/Auth/LoginReg';
-import Home from './components/Home/Home';
+import Home from './components/ShopNow/Home';
 import { useState } from 'react';
+import ProductDetails from './components/ShopNow/ProductDetails';
 
 
 
@@ -13,7 +14,7 @@ function App() {
   return (
     <div className="App">
       <Router>
-        
+
         <Routes>
           <Route
             path="/"
@@ -23,6 +24,12 @@ function App() {
             path="/home"
             element={isLoggedIn ? <Home setIsLoggedIn={setIsLoggedIn} /> : <Navigate to="/" />}
           />
+
+          <Route
+            path="/viewPrd/:prdId"
+            element={isLoggedIn ? <ProductDetails setIsLoggedIn={setIsLoggedIn} /> : <Navigate to="/" />}
+          />
+
 
         </Routes>
       </Router>
