--- src/components/Home/Home.css
@@ -1,7 +1,0 @@
-/* src/components/Home/Home.css */
-.home {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: space-around;
-    padding: 20px;
-}
