+++ src/components/ShopNow/NavBar.jsx
