+++ src/components/styles/LoginReg.css
@@ -10,7 +10,6 @@
 
 .shopping-container {
     /* added by vicky */
-    margin-top: 10%; 
     background: rgba(255, 255, 255, 0.9);
     padding: 30px;
     border-radius: 15px;
